Move code into utils/compareUtils.ts
